+++ allauth/account/adapter.py
@@ -676,6 +676,25 @@ class DefaultAccountAdapter(object):
     def authentication_failed(self, request, **credentials):
         pass
 
+    def reauthenticate(self, user, password):
+        from allauth.account.models import EmailAddress
+        from allauth.account.utils import user_email, user_username
+
+        credentials = {"password": password}
+        username = user_username(user)
+        if username:
+            credentials["username"] = username
+        email = None
+        primary = EmailAddress.objects.get_primary(user)
+        if primary:
+            email = primary.email
+        else:
+            email = user_email(user)
+        if email:
+            credentials["email"] = email
+        reauth_user = self.authenticate(context.request, **credentials)
+        return reauth_user is not None and reauth_user.pk == user.pk
+
     def is_ajax(self, request):
         return any(
             [